separate SV and CL steam functions
