+++ source/client/cl_mm.c
@@ -740,18 +740,18 @@ static bool CL_MM_LoginSteam( void )
 		cl_mm_steam_token = NULL;
 	}
 
-	if( Steam_GetAuthSessionTicket( &CL_MM_LoginSteamCb ) == 0 ) {
-		return false;
-	}
-
-	// wait for GetAuthSessionTicket callback but not for too long
-	start_time = Sys_Milliseconds();
-	while( !cl_mm_login_steam_complete ) { 
-		if( Sys_Milliseconds() > start_time + MM_STEAM_TICKET_AUTH_TIMEOUT ) {
-			break;
-		}
-		Steam_RunFrame(); 
-	}
+	// if( Steam_GetAuthSessionTicket( &CL_MM_LoginSteamCb ) == 0 ) {
+	// 	return false;
+	// }
+	//
+	// // wait for GetAuthSessionTicket callback but not for too long
+	// start_time = Sys_Milliseconds();
+	// while( !cl_mm_login_steam_complete ) { 
+	// 	if( Sys_Milliseconds() > start_time + MM_STEAM_TICKET_AUTH_TIMEOUT ) {
+	// 		break;
+	// 	}
+	// 	Steam_RunFrame(); 
+	// }
 	
 	return cl_mm_login_steam_ok;
 }
