+++ source/server/server.h
@@ -568,3 +568,13 @@ const char *SV_Web_UpstreamBaseUrl( void );
 bool SV_Web_AddGameClient( const char *session, int clientNum, const netadr_t *netAdr );
 void SV_Web_RemoveGameClient( const char *session );
 void SV_Web_GameFrame( http_game_query_cb cb );
+
+
+//
+// sv_steam.c
+//
+#include "../steamshim/src/steamshim_types.h"
+int Steam_GetAuthSessionTicket( void (*callback)( void *, size_t ) );
+int Steam_BeginAuthSession(uint64_t steamid, SteamAuthTicket_t *ticket);
+void Steam_EndAuthSession(uint64_t steamid);
+void SV_Steam_RunFrame( void );
